+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _This release is scheduled to be released on 2023-07-01._
 
 ### Added
 
+- Added UV Index to hourly and current Weather, with support for Openmeteo
 - Added tests for serveronly
 - Set Timezone `Europe/Berlin` in unit tests (needed for new formatTime tests)
 - Added no-param-reassign eslint rule and fix warnings